+++ api_test/testutil_test.go
@@ -215,6 +215,8 @@ func NewTestServer(t testing.TB, ctx context.Context, blocking bool) *TestServer
 	sqlOrderStore := sqlstore.NewOrders(&sqlStore)
 	sqlTradeStore := sqlstore.NewTrades(&sqlStore)
 	sqlNetworkLimitsStore := sqlstore.NewNetworkLimits(&sqlStore)
+	sqlAssetStore := sqlstore.NewAssets(&sqlStore)
+
 	eventSource, err := broker.NewEventSource(conf.Broker, logger)
 
 	if err != nil {
@@ -279,6 +281,7 @@ func NewTestServer(t testing.TB, ctx context.Context, blocking bool) *TestServer
 		sqlNetworkLimitsStore,
 		sqlMarketDataStore,
 		sqlTradeStore,
+		sqlAssetStore,
 	)
 	if srv == nil {
 		t.Fatal("failed to create gRPC server")